chore: fix errors
