+++ apps/academy-wep/app/page.tsx
@@ -1,10 +1,10 @@
 'use client'
 
+import AchievementsPage from '@/components/page/achivement'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { motion } from 'framer-motion'
 import { ArrowRight, DiscIcon as Discord, Github, Twitter } from 'lucide-react'
-import AchievementsPage from '@/components/Achivement'
 import { useState } from 'react'
 
 export default function Home() {
@@ -15,7 +15,6 @@ export default function Home() {
 	const now = new Date().getTime()
 	const distance = launchDate - now
 	const days = Math.floor(distance / (1000 * 60 * 60 * 24))
-
 
 	return (
 		<main className="min-h-screen relative overflow-hidden text-foreground">
@@ -136,7 +135,7 @@ export default function Home() {
 					))}
 				</motion.div>
 			</div>
-			<AchievementsPage/>
+			<AchievementsPage />
 		</main>
 	)
 }
