+++ apps/academy-wep/components/page/achivement.tsx
@@ -1,0 +1,81 @@
+import { Globe, Medal, Star, TrendingUp, Trophy, Wallet } from 'lucide-react'
+import type React from 'react'
+import AchievementCard, {
+	type AchievementCardProps,
+} from '../shared/achivement-card'
+
+const AchievementsPage: React.FC = () => {
+	const achievements: AchievementCardProps[] = [
+		{
+			title: 'Blockchain Basics',
+			description: 'Completed the Introduction to Blockchain module',
+			icon: Medal,
+			iconColor: '#10B981',
+			status: 'completed',
+			completedDate: 'April 10, 2025',
+			tokenId: 'GDJT4ENAXI4XBCX...',
+			viewLink: 'https://stellar.org',
+		},
+		{
+			title: 'Stellar Expert',
+			description: 'Mastered the Stellar blockchain fundamentals',
+			icon: Star,
+			iconColor: '#3B82F6',
+			status: 'locked',
+			moduleRequired: 'Stellar Blockchain Basics module required',
+		},
+		{
+			title: 'Wallet Master',
+			description: 'Learned how to manage Stellar wallets securely',
+			icon: Wallet,
+			iconColor: '#8B5CF6',
+			status: 'locked',
+			moduleRequired: 'Stellar Wallets module required',
+		},
+		{
+			title: 'Asset Manager',
+			description: 'Demonstrated proficiency in managing XLM and USDC',
+			icon: TrendingUp,
+			iconColor: '#F59E0B',
+			status: 'locked',
+			moduleRequired: 'Managing XLM & USDC module required',
+		},
+		{
+			title: 'Web3 Pioneer',
+			description: 'Connected projects to the broader web3 ecosystem',
+			icon: Globe,
+			iconColor: '#EF4444',
+			status: 'locked',
+			moduleRequired: 'Web3 Integration module required',
+		},
+		{
+			title: 'KindFi Graduate',
+			description: 'Completed all modules and demonstrated full proficiency',
+			icon: Trophy,
+			iconColor: '#06B6D4',
+			status: 'locked',
+			moduleRequired: 'All modules module required',
+		},
+	]
+
+	return (
+		<div className="bg-white min-h-screen p-8">
+			<div className=" mx-auto">
+				<h1 className="text-2xl font-bold mb-6 text-gray-800">
+					Your Achievement Collection
+				</h1>
+				<div className="grid grid-cols-1 md:grid-cols-3 gap-6">
+					{achievements.map((achievement, index) => (
+						<AchievementCard
+							// biome-ignore lint/suspicious/noArrayIndexKey: <explanation>
+							key={index}
+							{...achievement}
+						/>
+					))}
+				</div>
+			</div>
+		</div>
+	)
+}
+
+export default AchievementsPage
